+++ CHANGELOG.md
@@ -11,6 +11,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add visibility conditions to field container to simplify field display logic in field api. (#6024)
 - Use field size in bytes in field api to get precise result when compare field size. (#6024)
+- Return empty array if file does not exist in $_FILES in field api. (#6032)
+
+### New
+
+- Make field name required in field api. (#6032)
 
 ## 2.15.0 - 2021-10-11